Move attestation stats logic to explorer app.
