+++ apps/indexer/lib/indexer/celo/materialized_view_refresh.ex
@@ -46,29 +46,4 @@ defmodule Indexer.Fetcher.CeloMaterializedViewRefresh do
     end)
   end
 
-  def rebuild_attestation_stats(timeout) do
-    query = """
-      update celo_account
-      set celo_account.attestations_requested = stats.requested, celo_account.attestations_fulfilled = stats.fulfilled
-      where celo_account.address = stats.address
-      from (
-          select r.address, r.requested, f.fulfilled
-          from (
-              select celo_account.address, count(*) as requested
-              from logs, celo_account
-              where logs.first_topic='#{@attestation_issuer_selected}'
-              and logs.fourth_topic='0x000000000000000000000000'||encode(celo_account.address::bytea, 'hex') group by address
-          ) r
-          inner join (
-              select address, count(*) as fulfilled
-              from logs, celo_account
-              where first_topic='#{@attestation_completed}'
-              and fourth_topic='0x000000000000000000000000'||		encode(address::bytea, 'hex') group by address
-          ) f
-          on r.address = f.address
-      ) stats;
-    """
-
-    Repo.query!(query, [], timeout: timeout)
-  end
 end
